+++ example/android/app/src/main/kotlin/io/emma/emma_flutter_sample/MainActivity.kt
@@ -1,4 +1,4 @@
-package io.emma.emma_flutter_sdk_example
+package io.emma.emma_flutter_sample
 
 import io.flutter.embedding.android.FlutterActivity