+++ ios/Classes/SwiftEmmaFlutterSdkPlugin.swift
@@ -2,6 +2,123 @@ import Flutter
 import UIKit
 import EMMA_iOS
 
+extension FlutterAppDelegate : EMMAPushDelegate {
+    public func onPushOpen(_ push: EMMAPush) {
+        let _ = push.params
+        // treat params
+    }
+    
+    @objc
+    @available(iOS 10.0, *)
+    func userNotificationCenter(_ center: UNUserNotificationCenter,  willPresent notification: UNNotification, withCompletionHandler   completionHandler: @escaping (_ options:   UNNotificationPresentationOptions) -> Void) {
+        completionHandler([.badge, .sound])
+    }
+    
+    @objc
+    @available(iOS 10.0, *)
+    func userNotificationCenter(_ center: UNUserNotificationCenter, didReceive response: UNNotificationResponse, withCompletionHandler completionHandler: @escaping () -> Void) {
+        completionHandler()
+    }
+}
+
+/** This implementation replaces push methods on AppDelegate  */
+class EMMAFlutterPushDelegate {
+
+    let appDelegate = UIApplication.shared.delegate
+    
+    init() {
+        
+    }
+    
+    @objc
+    func application(_ application: UIApplication, didRegisterForRemoteNotificationsWithDeviceToken deviceToken: Data) {
+        EMMA.registerToken(deviceToken)
+        if appDelegate != nil && appDelegate!.responds(to: #selector(UIApplicationDelegate.application(_:didRegisterForRemoteNotificationsWithDeviceToken:))) {
+            appDelegate?.application?(application, didRegisterForRemoteNotificationsWithDeviceToken: deviceToken)
+        }
+    }
+    
+    @objc
+    func application(_ application: UIApplication, didReceiveRemoteNotification userInfo: [AnyHashable : Any], fetchCompletionHandler completionHandler: @escaping (UIBackgroundFetchResult) -> Void) {
+        EMMA.handlePush(userInfo)
+    }
+    
+    @objc
+    func application(_ application: UIApplication, didRegister notificationSettings: UIUserNotificationSettings) {
+        application.registerForRemoteNotifications()
+    }
+    
+    @objc
+    func application(_ application: UIApplication, didFailToRegisterForRemoteNotificationsWithError error: Error) {
+        NSLog("Error registering notifications " + error.localizedDescription);
+    }
+    
+    @objc
+    @available(iOS 10.0, *)
+    func userNotificationCenter(_ center: UNUserNotificationCenter,  willPresent notification: UNNotification, withCompletionHandler   completionHandler: @escaping (_ options:   UNNotificationPresentationOptions) -> Void) {
+        EMMA.handlePush(notification.request.content.userInfo)
+        completionHandler([.badge, .sound])
+    }
+    
+    @objc
+    @available(iOS 10.0, *)
+    func userNotificationCenter(_ center: UNUserNotificationCenter, didReceive response: UNNotificationResponse, withCompletionHandler completionHandler: @escaping () -> Void) {
+        EMMA.handlePush(response.notification.request.content.userInfo)
+        completionHandler()
+    }
+    
+    @available(iOS 10.0, *)
+    public func swizzlePushMethods() {
+        let appDelegate = UIApplication.shared.delegate
+        let appDelegateClass: AnyClass? = object_getClass(appDelegate)
+        
+        var swizzles = Array<(Selector, Selector)>()
+        
+        swizzles.append((#selector(FlutterAppDelegate.application(_:didRegisterForRemoteNotificationsWithDeviceToken:)),
+                         #selector(EMMAFlutterPushDelegate.self.application(_:didRegisterForRemoteNotificationsWithDeviceToken:))))
+        
+        swizzles.append((#selector(FlutterAppDelegate.application(_:didReceiveRemoteNotification:fetchCompletionHandler:)),
+                         #selector(EMMAFlutterPushDelegate.self.application(_:didReceiveRemoteNotification:fetchCompletionHandler:))))
+        
+        swizzles.append((#selector(FlutterAppDelegate.application(_:didRegister:)),
+                         #selector(EMMAFlutterPushDelegate.self.application(_:didRegister:))))
+        
+        swizzles.append((#selector(FlutterAppDelegate.application(_:didFailToRegisterForRemoteNotificationsWithError:)),
+                         #selector(EMMAFlutterPushDelegate.self.application(_:didFailToRegisterForRemoteNotificationsWithError:))))
+        
+        swizzles.append((#selector(FlutterAppDelegate.userNotificationCenter(_:willPresent:withCompletionHandler:)),
+                         #selector(EMMAFlutterPushDelegate.self.userNotificationCenter(_:willPresent:withCompletionHandler:))))
+        
+        swizzles.append((#selector(FlutterAppDelegate.userNotificationCenter(_:didReceive:withCompletionHandler:)),
+                         #selector(EMMAFlutterPushDelegate.self.userNotificationCenter(_:didReceive:withCompletionHandler:))))
+        
+        for s in swizzles {
+            
+            let originalSelector = s.0
+            let swizzledSelector = s.1
+            
+            guard let swizzledMethod = class_getInstanceMethod(EMMAFlutterPushDelegate.self, swizzledSelector) else {
+                return
+            }
+
+            if let originalMethod = class_getInstanceMethod(appDelegateClass, originalSelector)  {
+                method_exchangeImplementations(originalMethod, swizzledMethod)
+            } else {
+                class_addMethod(appDelegateClass, swizzledSelector, method_getImplementation(swizzledMethod), method_getTypeEncoding(swizzledMethod))
+            }
+        }
+
+        
+    }
+    
+}
+
+extension UIApplicationDelegate {
+    // MARK: - EMMA Push Delegate
+    
+
+}
+
 public class SwiftEmmaFlutterSdkPlugin: NSObject, FlutterPlugin {
   public static func register(with registrar: FlutterPluginRegistrar) {
     let channel = FlutterMethodChannel(name: "emma_flutter_sdk", binaryMessenger: registrar.messenger())
@@ -79,6 +196,7 @@ public class SwiftEmmaFlutterSdkPlugin: NSObject, FlutterPlugin {
             return
         }
         EMMA.trackExtraUserInfo(userAttributes)
+        result(nil)
         break
     case "loginUser":
         guard let args = call.arguments as? Dictionary<String, AnyObject> else {
@@ -98,6 +216,7 @@ public class SwiftEmmaFlutterSdkPlugin: NSObject, FlutterPlugin {
         let email = args["email"] as? String ?? ""
         
         EMMA.loginUser(userId, forMail: email)
+        result(nil)
         break
     case "registerUser":
         guard let args = call.arguments as? Dictionary<String, AnyObject> else {
@@ -117,6 +236,7 @@ public class SwiftEmmaFlutterSdkPlugin: NSObject, FlutterPlugin {
         let email = args["email"] as? String ?? ""
         
         EMMA.registerUser(userId, forMail: email)
+        result(nil)
         break
     case "inAppMessage":
         
@@ -143,7 +263,19 @@ public class SwiftEmmaFlutterSdkPlugin: NSObject, FlutterPlugin {
         
         let request = EMMAInAppRequest(type: requestType)
         EMMA.inAppMessage(request)
+        result(nil)
+        break
+    case "startPushSystem":
+        EMMA.startPushSystem()
         
+        if let applicationDelegate = UIApplication.shared.delegate as? FlutterAppDelegate {
+            let pushDelegate = EMMAFlutterPushDelegate()
+            if #available(iOS 10.0, *) {
+                pushDelegate.swizzlePushMethods()
+            }
+            EMMA.setPushSystemDelegate(applicationDelegate)
+        }
+        result(nil)
         break
       default:
         result(FlutterMethodNotImplemented)